Turned RequestContext into an extension class.
There should never be a reason to modify this class, so by turning it
into an extension class we gain increase performance and reduce the
memory footprint.

Also, this will allow us to store some private C attributes (like the
FUSE context object), which will be important for future extensions.