darwin_sem_wait: make local variable volatile
Otherwise we get an error from gcc because the 'res' variable
may get clobbered when "longjmp" or "vfork" is running. I am not sure
if this is a false positive or an actual bug -- it probably depends on
how pthread_cleanup_push/pop is implemented (apparently it uses
longjmp internally).