Wait for output on failed tests
If a test fails, wait a moment before retrieving the captured
stdout/stderr. When using a server process, this makes sure that we capture
any potential output of the server that comes *after* a test has failed. For
example, if a request handler raises an exception, the server first signals an
error to FUSE (causing the test to fail), and then logs the exception. Without
the extra delay, the exception will go into nowhere.