Do not use pthread_cancel()
Using pthread_cancel on worker threads can leave broken Python threads. In programs that create and destroy a llfuse mount many times (e.g. test suites), this can lead to hard-to-debug deadlocks later on. There is no safe way to use pthread_cancel, and it is unnecessary anyway because the worker threads will quit on their own when the kernel connection is closed.

Thanks to Peter Amstutz for the Patch.

Fixes issue #108.